Move integration tests to own file
